Manifestation sorting universal method

+++ src/main/java/org/ticketshop/repository/ManifestationRepository.java
@@ -1,7 +1,11 @@
 package org.ticketshop.repository;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.ticketshop.model.Manifestation;
 
-public interface ManifestationRepository extends JpaRepository<Manifestation, Integer> {
+import java.util.List;
+
+public interface ManifestationRepository extends JpaRepository<Manifestation, Long> {
+
 }
